+++ PolygonChecker/triangleAngleCalculator.c
@@ -1,4 +1,5 @@
-// Ryan Tu - November 2022 - software implementation group project
+// Ryan Tu - November 2022
+// Software Implementation group project
 // Calculates angles inside of triangle and returns sum
 
 #include <stdio.h>
@@ -8,15 +9,15 @@
 #define PI 3.14159265359
 
 
-float triangleAnglesCalculator(int* triangleSides, float* interiorAngles)
+double triangleAnglesCalculator(int* triangleSides, double* interiorAngles)
 {
-	float s1 = triangleSides[0];
-	float s2 = triangleSides[1];
-	float s3 = triangleSides[2];
+	double s1 = triangleSides[0];
+	double s2 = triangleSides[1];
+	double s3 = triangleSides[2];
 
 	// first angle 
-	float divideBy = (2.0*s1*s2);
-	float answerRadians = acos(((s1*s1) + (s2*s2) - (s3*s3)) / divideBy) ;
+	double divideBy = (2.0*s1*s2);
+	double answerRadians = acos(((s1*s1) + (s2*s2) - (s3*s3)) / divideBy) ;
 	interiorAngles[0] = ((answerRadians * 180.0) / PI);
 
 	// second angle
@@ -30,7 +31,7 @@ float triangleAnglesCalculator(int* triangleSides, float* interiorAngles)
 	interiorAngles[2] = ((answerRadians * 180.0) / PI);
 
 	// calculating the sum
-	float sumOfAngles = interiorAngles[0] + interiorAngles[1] + interiorAngles[2];
+	double sumOfAngles = interiorAngles[0] + interiorAngles[1] + interiorAngles[2];
 
 	return sumOfAngles;
